de-duplicate py_get_samples and periscope into rfmux.streamer.
There is still some duplicate code - but this is a start.
